+++ blocklets/ai-studio/src/pages/project/file-tree.tsx
@@ -15,14 +15,12 @@ import { css } from '@emotion/css';
 import { Icon } from '@iconify-icon/react';
 import ArrowBackUpIcon from '@iconify-icons/tabler/arrow-back-up';
 import ChevronDownIcon from '@iconify-icons/tabler/chevron-down';
-import CodeIcon from '@iconify-icons/tabler/code';
 import CopyIcon from '@iconify-icons/tabler/copy';
 import DotsVerticalIcon from '@iconify-icons/tabler/dots-vertical';
 import ExternalLinkIcon from '@iconify-icons/tabler/external-link';
 import FileDiscIcon from '@iconify-icons/tabler/file-description';
 import FolderPlusIcon from '@iconify-icons/tabler/folder-plus';
 import DiffIcon from '@iconify-icons/tabler/layers-difference';
-import LinkIcon from '@iconify-icons/tabler/link';
 import PencilIcon from '@iconify-icons/tabler/pencil';
 import TrashIcon from '@iconify-icons/tabler/trash';
 import { DragLayerMonitorProps, MultiBackend, NodeModel, Tree, getBackendOptions } from '@minoru/react-dnd-treeview';
@@ -502,26 +500,21 @@ const FileTree = forwardRef<
                     editing={meta.name === editingFileName}
                     selected={selected}
                     onClick={() => navigate(joinURL('.', filepath))}
-                    onDoubleClick={() => {
-                      if (!isEntryAgent) setEditingFileName(meta.name);
-                    }}
+                    onDoubleClick={() => setEditingFileName(meta.name)}
                     actions={actions}
                     sx={{ color: change?.color }}>
-                    {isEntryAgent ? (
-                      t('entryAgent')
-                    ) : (
-                      <EditTextItem
-                        editing={meta.name === editingFileName}
-                        onCancel={() => setEditingFileName(undefined)}
-                        onSubmit={async (name) => {
-                          setEditingFileName(undefined);
-                          const { data } = node;
-                          if (!data || name === data.name) return;
-                          meta.name = name;
-                        }}>
-                        {createFileName({ store, name: meta.name, defaultName: `${t('alert.unnamed')} Agent` })}
-                      </EditTextItem>
-                    )}
+                    <EditTextItem
+                      isEntry={isEntryAgent}
+                      editing={meta.name === editingFileName}
+                      onCancel={() => setEditingFileName(undefined)}
+                      onSubmit={async (name) => {
+                        setEditingFileName(undefined);
+                        const { data } = node;
+                        if (!data || name === data.name) return;
+                        meta.name = name;
+                      }}>
+                      {createFileName({ store, name: meta.name, defaultName: `${t('alert.unnamed')} Agent` })}
+                    </EditTextItem>
                   </TreeItem>
                   <AwarenessIndicator
                     projectId={projectId}
@@ -683,32 +676,12 @@ function TreeItemMenus({
         </MenuItem>
       ),
 
-      ...(item.type === 'folder' && onCreateFile
-        ? [
-            <MenuItem key="createPrompt" onClick={() => onCreateFile({ parent: item.path })}>
-              <ListItemIcon>
-                <Box component={Icon} icon={FileDiscIcon} />
-              </ListItemIcon>
-              <ListItemText primary={t('newObject', { object: t('prompt') })} />
-            </MenuItem>,
-
-            <MenuItem key="createApi" onClick={() => onCreateFile({ parent: item.path, meta: { type: 'api' } })}>
-              <ListItemIcon>
-                <Box component={Icon} icon={LinkIcon} />
-              </ListItemIcon>
-              <ListItemText primary={t('newObject', { object: t('api') })} />
-            </MenuItem>,
-
-            <MenuItem
-              key="createFunction"
-              onClick={() => onCreateFile({ parent: item.path, meta: { type: 'function' } })}>
-              <ListItemIcon>
-                <Box component={Icon} icon={CodeIcon} />
-              </ListItemIcon>
-              <ListItemText primary={t('newObject', { object: t('function') })} />
-            </MenuItem>,
-          ]
-        : []),
+      item.type === 'folder' && onCreateFile && (
+        <MenuItem key="createAgent" onClick={() => onCreateFile({ parent: item.path })}>
+          <ListItemIcon>{agentTypesMap.prompt?.icon} </ListItemIcon>
+          <ListItemText primary={t('newObject', { object: t('agent') })} />
+        </MenuItem>
+      ),
       item.type === 'file' && onCreateFile && (
         <MenuItem
           key="duplicateFile"
@@ -802,16 +775,19 @@ function TreeItemMenus({
 }
 
 function EditTextItem({
+  isEntry,
   editing,
   children,
   onCancel,
   onSubmit,
 }: {
+  isEntry?: boolean;
   editing?: boolean;
   children?: string;
   onCancel?: () => any;
   onSubmit?: (text: string) => any;
 }) {
+  const { t } = useLocaleContext();
   const [value, setValue] = useState(children);
 
   const submit = async () => {
@@ -863,7 +839,10 @@ function EditTextItem({
       onBlur={submit}
     />
   ) : (
-    children
+    <>
+      {children}
+      {isEntry && ` (${t('entryAgent')})`}
+    </>
   );
 }