+++ blocklets/ai-studio/src/pages/project/publish/publish-button.tsx
@@ -31,10 +31,10 @@ export default function PublishButton({ ...props }: LoadingButtonProps) {
       <Tooltip disableInteractive title={t('publish')}>
         <LoadingButton
           variant="outlined"
-          sx={{ minWidth: 0, minHeight: 0, height: 32, border: '1px solid #E5E7EB' }}
+          sx={{ px: 2, minWidth: 0, minHeight: 0, height: 32, border: '1px solid #E5E7EB' }}
           onClick={async (e) => {
             e.stopPropagation();
-            await saveButtonState.getState().save?.({ skipConfirm: true });
+            await saveButtonState.getState().save?.({ skipConfirm: true, skipCommitIfNoChanges: true });
             setShowCreateResource(true);
           }}
           startIcon={<Box component={Icon} icon={BrandAppgalleryIcon} />}
@@ -51,7 +51,7 @@ export default function PublishButton({ ...props }: LoadingButtonProps) {
 }
 
 function PublishDialog({ project, onClose }: { project: Project; onClose: () => void }) {
-  const [logo] = useState(() => getProjectIconUrl(project._id, project.updatedAt, { original: true }));
+  const [logo] = useState(() => getProjectIconUrl(project.id, project.updatedAt, { original: true }));
   const [opened, setOpened] = useState(false);
 
   if (!project) return null;
@@ -61,7 +61,7 @@ function PublishDialog({ project, onClose }: { project: Project; onClose: () => 
       <BlockletStudio
         style={{ opacity: opened ? 1 : 0 }}
         mode="dialog"
-        tenantScope={project._id}
+        tenantScope={project.id}
         title={project.name || ''}
         description={project.description || ''}
         note=""
@@ -69,7 +69,7 @@ function PublishDialog({ project, onClose }: { project: Project; onClose: () => 
         logo={logo}
         componentDid={AI_STUDIO_COMPONENT_DID}
         // 透传到 get blocklet resource 的参数
-        resourcesParams={{ projectId: project._id }}
+        resourcesParams={{ projectId: project.id }}
         dependentComponentsMode="readonly"
         open
         setOpen={() => onClose()}